+++ CHANGELOG.md
@@ -15,7 +15,15 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
 
 
+## 0.70.2
+
+### 🐛 Fixes
+- [8045](https://github.com/vegaprotocol/vega/issues/8045) - Fix bug in handling internal sources data.
+
+
 ## 0.70.1
+- [7943](https://github.com/vegaprotocol/vega/issues/7943) - Add version to network file to be future-proof.
+- [8045](https://github.com/vegaprotocol/vega/issues/8045) - Fix bug in handling internal sources data.
 
 ### 🐛 Fixes
 - [7950](https://github.com/vegaprotocol/vega/issues/7950) - Fix the restore of deposits from checkpoint